Add support for parsing sfacct records
Classic usage is centred around pmacct on fishhook, which has a
particular output format (and is an older version of the software).
We're running sfacct on acct1.lax1 instead, and there's many more output
fields in its version that we need to parse, so we have to deal with
that case too.

I've torn up the old way of parsing args and replaced it with getopt.
Not the most elegant thing, but it'll get the job done.

At the moment the expected input format is fixed. It's not ideal in case
the field ordering changes again in future, but we're probably not *too*
worried about that for now.